Renames ConversationList as UserList
